pyproject.toml
[build-system]
requires = ["scikit-build-core>=0.4.3", "nanobind>=2.11.0"]
build-backend = "scikit_build_core.build"

[project]
name = "pytalib"
version = "0.1.0"
description = "Python wrapper for ta-lib using nanobind"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy",
    "pandas"
]

[tool.scikit-build]
# Protect the configuration against future changes in scikit-build-core
minimum-version = "0.4"
# Set CMake build type
cmake.build-type = "Release"

[dependency-groups]
dev = [
    "ta-lib>=0.6.8",
    "pytest-benchmark>=4.0.0",
]
